Added filtration before adding player to event.

+++ GoSportDB/routes/events/events-controller.js
@@ -167,7 +167,7 @@ const controller = {
                 event.admin.username + " Ви покани да се присъедините", playersTokens);
         });
     },
-    addUserToEvent(req, res, eventRepository, userRepository) {
+    addUserToEvent(req, res, eventRepository, userRepository, teamRepository) {
         const userId = +req.body.userId;
         const eventId = +req.params.id;
 
@@ -198,33 +198,51 @@ const controller = {
                 }
 
                 const event = events[0];
-                if (event.neededPlayers == -1 ||
-                    event.players.length < event.neededPlayers) {
-                    for (let i = 0; i < event.players.length; i += 1) {
-                        if (event.players[i].id == userId) {
-                            res.send('Already in the event');
-                            return;
-                        }
-                    }
-                    event.players.push(mappedUser);
-                    eventRepository.removeEvent(eventId).then(() => {
-                        eventRepository.insertEvent(event).then(() => {
-                            user.events.push(event);
-                            userRepository.removeUser(user.id).then(() => {
-                                userRepository.insertUser(user).then(() => {
-                                    this.notifyPlayersForNewPlayer(event, user);
-                                    res.send(event);
-                                    return;
-                                });
-                            });
-                        });
-                    });
+                if (!event.teamIds || event.teamIds.length === 0) {
+                    this._addPlayerToEvent(req, res, userRepository, eventRepository, mappedUser, event, null);
                 } else {
-                    res.send('Maximum players reached');
+                    teamRepository.getAllTeams().then((teams) => {
+                        const playerIds = [].concat((teams.filter((t) => {
+                            return event.teamIds.find((i) => i === t.id);
+                        })).map((t) => t.players)).map((p) => p.id);
+                        // Calls with parameter for ids of players who can participate
+                        this._addPlayerToEvent(req, res, userRepository, eventRepository, mappedUser, event, playerIds);
+                    });
                 }
             });
         });
 
+    },
+
+    _addPlayerToEvent(req, res, userRepository, eventRepository, mappedUser, event, playerIds) {
+        if (playerIds && (playerIds.indexOf(mappedUser.id) < 0)) {
+            res.send('Събитието е ограничено само за някои отбори');
+            return;
+        }
+        if (event.neededPlayers == -1 ||
+            event.players.length < event.neededPlayers) {
+            for (let i = 0; i < event.players.length; i += 1) {
+                if (event.players[i].id == userId) {
+                    res.send('Already in the event');
+                    return;
+                }
+            }
+            event.players.push(mappedUser);
+            eventRepository.removeEvent(eventId).then(() => {
+                eventRepository.insertEvent(event).then(() => {
+                    user.events.push(event);
+                    userRepository.removeUser(user.id).then(() => {
+                        userRepository.insertUser(user).then(() => {
+                            this.notifyPlayersForNewPlayer(event, user);
+                            res.send(event);
+                            return;
+                        });
+                    });
+                });
+            });
+        } else {
+            res.send('Няма повече ограничени места');
+        }
     },
     notifyPlayersForNewPlayer(event, player) {
         const mappedTokens = event.players.map((m) => m.token)
